Simplify download_pmhc_mds function
Previously, we tried to be clever and download cached versions of files
extracted with the same parameters. However, this complicated the code,
and in common usage, a script is going to want a fresh extract. Now, we
just request a new extract, wait patiently for it to finish, then
download it.

I've also added some additional arguments to specify more PMHC MDS
extract options.