Retry after timeout when waiting for PMHC extract
For large extract, we have to wait a long time for the extract to
generate. While waiting, we make a request every 30 seconds to check if
the extract is ready. There are a lot of chances for request timeouts
during this period. This commit adds a max_retries parameter, which
will retry up to the specified number of timeouts. By default, this is
set to 20 retries.