Don't assume that PMHC extracts will succeed
Previously, we assumed that all PMHC extracts would eventually change
from status=Processing to status=Completed, at which point we could
download the extract file. However, it is possible that the PMHC server
may error, which results in status=Error. We did not detect this Error
state previously, so would loop indefinitely in the false hope that the
extract file would eventually be created.

Now, we correctly detect status=Error, and raise a suitable exception.